implementacion reporte y pruebas

+++ microservicios-spring/cuenta/src/main/java/com/ms/cliente/cuenta/dto/CuentaDTO.java
@@ -3,6 +3,7 @@ package com.ms.cliente.cuenta.dto;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -12,7 +13,7 @@ public class CuentaDTO {
     private String id;
     private String numeroCuenta;
     private String tipoCuenta;
-    private double saldoInicial;
+    private BigDecimal saldoInicial;
     private String estado;
     private String clienteId;
     private List<MovimientoDTO> movimientos;
