+++ microservicios-spring/cuenta/src/main/java/com/ms/cliente/cuenta/dto/MovimientoDTO.java
@@ -3,6 +3,7 @@ package com.ms.cliente.cuenta.dto;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -11,7 +12,7 @@ public class MovimientoDTO {
     private String id;
     private Date fecha;
     private String tipoMovimiento;
-    private double valor;
-    private double saldo;
+    private BigDecimal valor;
+    private BigDecimal saldo;
 
 }
