+++ microservicios-spring/cuenta/src/main/java/com/ms/cliente/cuenta/model/Cuenta.java
@@ -6,6 +6,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,7 +17,7 @@ public class Cuenta {
     private String id;
     private String numeroCuenta;
     private String tipoCuenta;
-    private double saldoInicial;
+    private BigDecimal saldoInicial;
     private String estado;
     private String clienteId;
     @DBRef
