+++ microservicios-spring/cuenta/src/main/java/com/ms/cliente/cuenta/model/Movimiento.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -16,7 +17,7 @@ public class Movimiento {
     private String id;
     private Date fecha;
     private String tipoMovimiento;
-    private double valor;
-    private double saldo;
+    private BigDecimal valor;
+    private BigDecimal saldo;
 
 }
